backend/mis/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2
django-cors==0.1
django-cors-headers==4.3.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
importlib-metadata==7.0.1
Markdown==3.5.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
zipp==3.17.0
